Create ADR 0001: Record architecture decisions
The content of this ADR is based heavily on the LAA Hosting equivalent ADR
[1], which in turn is based on Michael Nygard's post about ADRs [2].

I've marked this ADR as Accepted because we've already agreed to use ADRs. The
details of how we use them are still open for discussion, which can happen on
the pull request.

Refer to this ADR from the README.

Include adr-tools config to specify the directory to keep ADRs in, matching
the decision recorded in 0001.

[1]: https://github.com/ministryofjustice/laa-hosting-architectural-decisions/blob/827b8832c32577119f8ac49e4520fc689a3bded1/doc/adr/0001-record-architecture-decisions.md
[2]: http://thinkrelevance.com/blog/2011/11/15/documenting-architecture-decisions