Add a table of contents to the README
This was generated with:

    adr generate toc -p decisions/

adr-tools writes this to stdout and doesn't support inserting it into a file,
so I've copied it across manually for now. We could script the generation of
the README including this output in future.

adr-log also exists and can insert a table of contents into an existing file,
but it's an npm package and I'd rather not introduce that as a dependency in
the context of this repo.