Add ADR 0013 on using NOMIS SSO for user auth
I've backdated this ADR to the date when we made the decision because the
context has changed a bit since then, and I want to avoid retrospective
confusion about the timing of this decision and others which followed.

Since we made this decision:

- we integrated the allocation manager with NOMIS SSO
- we identified a way of getting access to the Delius API much earlier than we
  expected, which makes it more important that we get our applications into
  production quickly to enable that work
- we've started building our production environment
- NOMIS SSO has been released to production

The decision we made on this is still in line with these developments, but I'm
documenting this to make it clear that the decision was made at a point in
time when we didn't expect these things to happen this quickly.