Add diagram of build & deployment pipeline
This covers everything which our team is responsible for (in colour) as well
as the external HMPPS services we depend on.

I started making this diagram to get some security-related things straight in
my head, which is why the Kubernetes secrets and service accounts are
included. Other Kubernetes resources (or how those managed as part of cloud
platform environments are deployed) are not shown.

Everything in this diagram exists now except for our connection to the Delius
API and the corresponding staging setup, our plans for which are documented in
ADR 0018.