Unify API and Presentation layers
This PR discusses the decision reached with the team around simplifying
development, deployment and management of the allocation tool.  This is
essentially moving to a monolith as a starting point, until we have an
operational need to break the service apart, at which point we can
define the boundaries based on experience of running the system.